messaging_app/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
